rename filter_event in numpy to filter
