+++ temporian/implementation/numpy/operators/all_operators.py
@@ -1,7 +1,8 @@
 """Imports all the packages containing operator implementations."""
 
 # pylint: disable=unused-import
-from temporian.implementation.numpy.operators import filter_event
+# pylint: disable=redefined-builtin
+from temporian.implementation.numpy.operators import filter
 from temporian.implementation.numpy.operators import glue
 from temporian.implementation.numpy.operators import lag
 from temporian.implementation.numpy.operators import prefix
@@ -27,3 +28,4 @@ from temporian.implementation.numpy.operators.calendar import day_of_month
 from temporian.implementation.numpy.operators.calendar import day_of_week
 
 # pylint: enable=unused-import
+# pylint: enable=redefined-builtin
