+++ temporian/implementation/numpy/operators/filter.py
